+++ stac-cli/src/commands/mod.rs
@@ -1,5 +1,6 @@
+mod item;
 mod search;
 mod sort;
 mod validate;
 
-pub use {search::search, sort::sort, validate::validate};
+pub use {item::item, search::search, sort::sort, validate::validate};